cfr util
